Add files via upload
In Power BI, you can create visualization in two ways. First is by adding from the right side
pane to Report Canvas. By default, it is the table type visualization, which is selected in
Power BI. Another way is to drag the fields from right side bar to the axis and value axis under
Visualization.
In Power BI, it is also possible to move your visualization on the reporting canvas by clicking
and then dragging it. You can also switch between different type of charts and visualizations
from the Visualization pane. Power BI attempts to convert your selected fields to the new
visual type as closely as possible.